refactored

+++ src/database/curd.go
@@ -1,10 +1,14 @@
 package database
 
 import (
+	"github.com/umangshrestha/todo-app/src/config"
 	"gorm.io/gorm"
 )
 
 func CreateTodo(db *gorm.DB, todo *Todo) (uint, error) {
+	if db == nil {
+		return 0, ErrDBNil
+	}
 	result := db.Create(todo)
 	if result.Error != nil {
 		return 0, result.Error
@@ -12,17 +16,23 @@ func CreateTodo(db *gorm.DB, todo *Todo) (uint, error) {
 	return todo.ID, nil
 }
 
-func FindAll(db *gorm.DB, limit, offset int, isDeleted, isCompleted bool) ([]*Todo, error) {
+func FindAllTodo(db *gorm.DB, q *Query) ([]*Todo, error) {
+	if db == nil {
+		return nil, ErrDBNil
+	}
 	var todos []*Todo
-	query := db.Order("created_at DESC").Limit(limit).Offset(offset)
+	if q.Limit == 0 {
+		q.Limit = config.Limit
+	}
+	query := db.Order("created_at DESC").Limit(q.Limit).Offset(q.Offset)
 
-	if isDeleted {
+	if q.Deleted {
 		query = query.Where("deleted_at IS NOT NULL")
 	} else {
 		query = query.Where("deleted_at IS NULL")
 	}
 
-	if isCompleted {
+	if q.Completed {
 		query = query.Where("completed_at IS NOT NULL")
 	} else {
 		query = query.Where("completed_at IS NULL")
@@ -37,7 +47,7 @@ func FindAll(db *gorm.DB, limit, offset int, isDeleted, isCompleted bool) ([]*To
 	return todos, nil
 }
 
-func FindById(db *gorm.DB, id uint) (*Todo, error) {
+func FindTodoById(db *gorm.DB, id uint) (*Todo, error) {
 	var todo Todo
 	result := db.First(&todo, id)
 	if result.Error != nil {
@@ -47,6 +57,9 @@ func FindById(db *gorm.DB, id uint) (*Todo, error) {
 }
 
 func UpdateTodoById(db *gorm.DB, id uint, updates *Todo) error {
+	if db == nil {
+		return ErrDBNil
+	}
 	result := db.Model(&Todo{}).Where("id = ?", id).Updates(updates)
 	if result.Error != nil {
 		return result.Error
@@ -55,9 +68,24 @@ func UpdateTodoById(db *gorm.DB, id uint, updates *Todo) error {
 }
 
 func DeleteTodoById(db *gorm.DB, id uint) error {
+	if db == nil {
+		return ErrDBNil
+	}
 	result := db.Delete(&Todo{}, id)
 	if result.Error != nil {
 		return result.Error
 	}
 	return nil
+}
+
+func CountTodo(db *gorm.DB) (*Count, error) {
+	if db == nil {
+		return nil, ErrDBNil
+	}
+	var count Count
+	db.Model(&Todo{}).Where("completed_at IS NOT NULL").Where("deleted_at IS NOT NULL").Count(&count.Completed)
+	db.Model(&Todo{}).Where("deleted_at IS NULL").Where("completed_at IS NULL").Count(&count.Todo)
+	db.Model(&Todo{}).Where("deleted_at IS NOT NULL").Count(&count.Deleted)
+	db.Model(&Todo{}).Count(&count.Total)
+	return &count, nil
 }
